prebuilt: Remove Chromium app
Remove prebuilt Chromium app. I don’t want to store it here
because of its large size. Moreover, some libraries must be
extracted and placed in the /system/lib/ for the application
work correctly, therefore we lose 100+ Mb of free space on
the system partition of the device.

Sorry Chromium, but we should say goodbye to you.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>